#3149 - improved abstraction to remove code duplicity, improved FE behavior
